Shipment scan

+++ app/src/main/java/org/getcarebase/carebase/viewmodels/DeviceViewModel.java
@@ -10,7 +10,6 @@ import androidx.lifecycle.ViewModel;
 import org.getcarebase.carebase.R;
 import org.getcarebase.carebase.models.DeviceModel;
 import org.getcarebase.carebase.models.PendingDevice;
-import org.getcarebase.carebase.models.Procedure;
 import org.getcarebase.carebase.models.Shipment;
 import org.getcarebase.carebase.models.User;
 import org.getcarebase.carebase.repositories.DeviceRepository;
@@ -59,9 +58,15 @@ public class DeviceViewModel extends ViewModel {
 
     private final SingleEventMediatorLiveData<List<Shipment>> shipmentsLiveData = new SingleEventMediatorLiveData<>();
 
+    private final SingleEventMediatorLiveData<Shipment> shipmentLiveData = new SingleEventMediatorLiveData<>();
+
     private final MutableLiveData<Shipment> saveShipmentLiveData = new MutableLiveData<>();
     private final LiveData<Request> saveShipmentRequestLiveData =
             Transformations.switchMap(saveShipmentLiveData, shipment -> shipmentRepository.saveShipment(shipment));
+
+    private final MutableLiveData<Shipment> receiveShipmentLiveData = new MutableLiveData<>();
+    private final LiveData<Request> receiveShipmentRequestLiveData =
+            Transformations.switchMap(receiveShipmentLiveData,shipment -> shipmentRepository.receiveShipment(shipment));
 
     public DeviceViewModel() {
         authRepository = new FirebaseAuthRepository();
@@ -70,7 +75,9 @@ public class DeviceViewModel extends ViewModel {
     @Override
     protected void onCleared() {
         super.onCleared();
-        deviceRepository.destroy();
+        if (deviceRepository != null) {
+            deviceRepository.destroy();
+        }
     }
 
     public LiveData<Resource<User>> getUserLiveData() {
@@ -121,6 +128,11 @@ public class DeviceViewModel extends ViewModel {
         return shipmentsLiveData.getLiveData();
     }
 
+    public LiveData<Resource<Shipment>> getShipment(String shipmentId) {
+        shipmentLiveData.addSource(shipmentRepository.getShipment(shipmentId));
+        return shipmentLiveData.getLiveData();
+    }
+
     public void savePhysicalLocation(String physicalLocation) {
         savePhysicalLocationLiveData.setValue(physicalLocation);
     }
@@ -145,6 +157,16 @@ public class DeviceViewModel extends ViewModel {
         saveShipmentLiveData.setValue(shipment);
     }
 
+    public void receiveShipment(List<Map<String,String>> items) {
+        Shipment shipment = Objects.requireNonNull(shipmentLiveData.getLiveData().getValue()).getData();
+        shipment.setItems(items);
+        receiveShipmentLiveData.setValue(shipment);
+    }
+
+    public LiveData<Request> getReceiveShipmentRequestLiveData() {
+        return receiveShipmentRequestLiveData;
+    }
+
     public LiveData<Resource<DeviceModel>> getAutoPopulatedDeviceLiveData() {
         return autoPopulatedDeviceLiveData;
     }
@@ -167,43 +189,38 @@ public class DeviceViewModel extends ViewModel {
 
     public void autoPopulatedScannedBarcode(String barcode) {
         autoPopulatedDeviceLiveData.setValue(new Resource<>(null,new Request(null,Request.Status.LOADING)));
-        List<LiveData<Resource<DeviceModel>>> databaseSources = deviceRepository.autoPopulateFromDatabaseAndShipment(barcode);
-        LiveData<Resource<DeviceModel>> inventorySource = databaseSources.get(0);
-        LiveData<Resource<DeviceModel>> shippedSource = databaseSources.get(1);
+        LiveData<Resource<DeviceModel>> inventorySource = deviceRepository.autoPopulateFromDatabase(barcode);
         LiveData<Resource<DeviceModel>> gudidSource = deviceRepository.autoPopulateFromGUDID(barcode);
-        DeviceSourceObserver deviceSourceObserver = new DeviceSourceObserver(inventorySource,shippedSource,gudidSource);
+        DeviceSourceObserver deviceSourceObserver = new DeviceSourceObserver(inventorySource,gudidSource);
         autoPopulatedDeviceLiveData.addSource(inventorySource,deviceSourceObserver);
-        autoPopulatedDeviceLiveData.addSource(shippedSource,deviceSourceObserver);
         autoPopulatedDeviceLiveData.addSource(gudidSource,deviceSourceObserver);
     }
 
     private class DeviceSourceObserver implements Observer<Resource<DeviceModel>> {
         private final LiveData<Resource<DeviceModel>> inventorySource;
-        private final LiveData<Resource<DeviceModel>> shippedSource;
         private final LiveData<Resource<DeviceModel>> gudidSource;
-        public DeviceSourceObserver(LiveData<Resource<DeviceModel>> inventorySource, LiveData<Resource<DeviceModel>> shippedSource, LiveData<Resource<DeviceModel>> gudidSource) {
+
+        public DeviceSourceObserver(LiveData<Resource<DeviceModel>> inventorySource,LiveData<Resource<DeviceModel>> gudidSource) {
             this.inventorySource = inventorySource;
-            this.shippedSource = shippedSource;
             this.gudidSource = gudidSource;
         }
         @Override
         public void onChanged(Resource<DeviceModel> deviceModelResource) {
             if (deviceModelResource.getRequest().getStatus() == Request.Status.SUCCESS
                 || deviceModelResource.getRequest().getStatus() == Request.Status.ERROR) {
-                mediateDataSource(inventorySource,shippedSource,gudidSource);
+                mediateDataSource(inventorySource,gudidSource);
             }
         }
     }
 
-    private void mediateDataSource(LiveData<Resource<DeviceModel>> inventorySource, LiveData<Resource<DeviceModel>> shippedSource, LiveData<Resource<DeviceModel>> gudidSource) {
+    private void mediateDataSource(LiveData<Resource<DeviceModel>> inventorySource, LiveData<Resource<DeviceModel>> gudidSource) {
         Resource<DeviceModel> inventoryResource = Objects.requireNonNull(inventorySource.getValue());
-        Resource<DeviceModel> shippedResource = Objects.requireNonNull(shippedSource.getValue());
         Resource<DeviceModel> gudidResource = Objects.requireNonNull(gudidSource.getValue());
-        if (inventoryResource.getRequest().getStatus() == Request.Status.LOADING || shippedResource.getRequest().getStatus() == Request.Status.LOADING ||  gudidResource.getRequest().getStatus() == Request.Status.LOADING) {
+        if (inventoryResource.getRequest().getStatus() == Request.Status.LOADING ||  gudidResource.getRequest().getStatus() == Request.Status.LOADING) {
             return;
         }
 
-        boolean databaseError = inventoryResource.getRequest().getStatus() == Request.Status.ERROR && shippedResource.getRequest().getStatus() == Request.Status.ERROR;
+        boolean databaseError = inventoryResource.getRequest().getStatus() == Request.Status.ERROR;
         if (gudidResource.getRequest().getStatus() == Request.Status.ERROR && gudidResource.getRequest().getResourceString() == R.string.error_something_wrong) {
             autoPopulatedDeviceLiveData.setValue(gudidResource);
         }
@@ -228,24 +245,11 @@ public class DeviceViewModel extends ViewModel {
             // device that is not in gudid but has device model information in database
             // device could not be auto populated (no data)
             // or device is in our database
-            if (shippedResource.getRequest().getStatus() == Request.Status.SUCCESS) {
-                if (inventoryResource.getRequest().getResourceString() != null && inventoryResource.getRequest().getResourceString() == R.string.error_device_lookup) {
-                    shippedResource.getData().getProductions().get(0).setPhysicalLocation("");
-                    shippedResource.getData().setEquipmentType("");
-                    shippedResource.getData().setQuantity(0);
-                    autoPopulatedDeviceLiveData.setValue(shippedResource);
-                } else {
-                    inventoryResource.getData().setShipment(shippedResource.getData().getShipment());
-                    autoPopulatedDeviceLiveData.setValue(inventoryResource);
-                }
-            } else {
-                autoPopulatedDeviceLiveData.setValue(inventoryResource);
-            }
+            autoPopulatedDeviceLiveData.setValue(inventoryResource);
         }
 
         // stop listening to these sources
         autoPopulatedDeviceLiveData.removeSource(inventorySource);
-        autoPopulatedDeviceLiveData.removeSource(shippedSource);
         autoPopulatedDeviceLiveData.removeSource(gudidSource);
     }
 }
